Finished the front end CRUD operations (#1)
