+++ src/client-app/src/common/enums/availableProductActions.ts
@@ -1,0 +1,7 @@
+enum AvailableProductActions {
+    SetId,
+    SetName,
+    SetPrice
+};
+
+export default AvailableProductActions;
